pyproject.toml
[project]
name = "final"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "concrete-python>=2.10.0",
    "flask>=3.1.1",
    "setuptools>=80.9.0",
    "wheel>=0.45.1",
]
